+++ packages/voidjs-cli/src/webpack-plugins/HtmlPluginForDevServer.ts
